requirements.txt
Package         Version
--------------- -----------
cryptography    42.0.7
matplotlib      3.8.3
numpy           1.26.4
pillow          10.2.0
pip             22.3
pycryptodome    3.20.0
tqdm            4.66.4
